chore: deduplicate code for Kubernetes types
